Finish the product documentation and fixed any database logic

+++ Backend/MunchEZ/src/main/resources/db/migration/V1__create-table-products.sql
@@ -1,7 +1,7 @@
 CREATE TABLE products (
     id INT NOT NULL AUTO_INCREMENT,
     name VARCHAR(100) NOT NULL,
-    price INT NOT NULL,
+    price DECIMAL(10,2) NOT NULL,
     description VARCHAR(200),
     type ENUM('FOOD', 'DRINKS', 'DESSERTS'),
     active TINYINT DEFAULT 1,
